requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
aiofiles==24.1.0
    # via vs-project (pyproject.toml)
aiosqlite==0.21.0
    # via vs-project (pyproject.toml)
alembic==1.15.2
    # via vs-project (pyproject.toml)
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via
    #   httpx
    #   starlette
    #   watchfiles
asyncpg==0.30.0
    # via vs-project (pyproject.toml)
black==25.1.0
    # via vs-project (pyproject.toml)
certifi==2025.1.31
    # via
    #   httpcore
    #   httpx
    #   requests
cfgv==3.4.0
    # via pre-commit
charset-normalizer==3.4.1
    # via requests
click==8.1.8
    # via
    #   black
    #   rich-toolkit
    #   typer
    #   uvicorn
distlib==0.3.9
    # via virtualenv
dnspython==2.7.0
    # via email-validator
email-validator==2.2.0
    # via fastapi
fastapi==0.115.12
    # via vs-project (pyproject.toml)
fastapi-cli==0.0.7
    # via fastapi
filelock==3.18.0
    # via virtualenv
greenlet==3.2.1
    # via vs-project (pyproject.toml)
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.8
    # via httpx
httptools==0.6.4
    # via uvicorn
httpx==0.28.1
    # via fastapi
identify==2.6.10
    # via pre-commit
idna==3.10
    # via
    #   anyio
    #   email-validator
    #   httpx
    #   requests
iniconfig==2.1.0
    # via pytest
jinja2==3.1.6
    # via fastapi
mako==1.3.10
    # via alembic
markdown-it-py==3.0.0
    # via rich
markupsafe==3.0.2
    # via
    #   jinja2
    #   mako
mdurl==0.1.2
    # via markdown-it-py
mypy-extensions==1.1.0
    # via black
nodeenv==1.9.1
    # via pre-commit
packaging==25.0
    # via
    #   black
    #   pytest
pathspec==0.12.1
    # via black
platformdirs==4.3.7
    # via
    #   black
    #   virtualenv
pluggy==1.5.0
    # via pytest
pre-commit==4.2.0
    # via vs-project (pyproject.toml)
pydantic==2.11.3
    # via
    #   vs-project (pyproject.toml)
    #   fastapi
pydantic-core==2.33.1
    # via pydantic
pygments==2.19.1
    # via rich
pytest==8.3.5
    # via vs-project (pyproject.toml)
python-dotenv==1.1.0
    # via uvicorn
python-multipart==0.0.20
    # via fastapi
pyyaml==6.0.2
    # via
    #   pre-commit
    #   uvicorn
redis==5.2.1
    # via vs-project (pyproject.toml)
requests==2.32.3
    # via vs-project (pyproject.toml)
rich==14.0.0
    # via
    #   rich-toolkit
    #   typer
rich-toolkit==0.14.3
    # via fastapi-cli
shellingham==1.5.4
    # via typer
sniffio==1.3.1
    # via anyio
sqlalchemy==2.0.40
    # via
    #   vs-project (pyproject.toml)
    #   alembic
starlette==0.46.2
    # via fastapi
typer==0.15.2
    # via fastapi-cli
typing-extensions==4.13.2
    # via
    #   aiosqlite
    #   alembic
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   rich-toolkit
    #   sqlalchemy
    #   typer
    #   typing-inspection
typing-inspection==0.4.0
    # via pydantic
urllib3==2.4.0
    # via requests
uvicorn==0.34.2
    # via
    #   vs-project (pyproject.toml)
    #   fastapi
    #   fastapi-cli
uvloop==0.21.0
    # via uvicorn
virtualenv==20.30.0
    # via pre-commit
watchfiles==1.0.5
    # via uvicorn
websockets==15.0.1
    # via uvicorn
psycopg2-binary